Add files via upload
The Robo car is controlled by an Arduino Nano and powered by two 3V DC-Motors the ML(left) and the MR(right). The movement of the car is controlled by a IR-Receiver but the problem is that the arduino only execute the first command which is made with the controller and stays in this command.